Add files via upload
 I have successfully completed my first task as a Data Analyst Intern at Future Intern!

During this project, I focused on handling missing values and outliers, which are critical steps in data preprocessing. 

Importance of Data Cleaning: Ensuring data quality is paramount. Handling missing values appropriately ensures that our models are built on reliable data, leading to more accurate and trustworthy results.

Techniques for Managing Missing Values:
Imputation: Replacing missing values with mean, median, or mode.
Deletion: Removing rows or columns with missing values when appropriate.
Prediction Models: Using algorithms to predict and fill in missing data.
Dealing with Outliers:

Detection and Handling with IQR: Used the Interquartile Range (IQR) method to detect and handle outliers. This involved identifying values that fall significantly outside the expected range and deciding whether to remove or transform them to minimize their impact on the analysis.
Tools and Technologies: Leveraged Python and libraries like Pandas, NumPy, and Matplotlib to effectively manage and visualize the data.

Practical Insights: This project highlighted the importance of a meticulous approach to data preprocessing, setting a solid foundation for subsequent analysis and modeling.
#Future Intern